### 2019-08-15 #1648 @whedon
[PRE REVIEW]: 3D reconstruction toolbox for behavior tracked with multiple cameras
**Submitting author:** @SwathiSheshadri (<a href="http://orcid.org/0000-0003-2850-107X">Swathi Sheshadri</a>)
**Repository:** <a href="https://github.com/SwathiSheshadri/pose3d" target ="_blank">https://github.com/SwathiSheshadri/pose3d</a>
**Version:** v1.0.0
