### 2019-08-15 #1648 @danielskatz
👋 @SwathiSheshadri - as you can see in the PDF, there are a few problems with the paper, such as references without a space before them, and a missing References header for the reference section.  Please feel free to make changes in the .md file, then put a comment ```@whedon generate pdf``` in this issue to regenerate the pdf
