+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Markerless tracking is a crucial experimental requirement for behavioral studies conducted in many species in different environments. A recently developed toolbox called DeepLabCut (DLC) (@Mathis2018) leverages Artificial Neural Network (ANN) based computer vision to make precise markerless tracking possible for scientific experiments. DLC uses a deep convolutional neural network, ResNet(@He2016) pre-trained on ImageNet database(@Imagenet) and adapts it to make it applicable for behavioral tracking tasks. To track complex behaviors such as grasping with object interaction in 3D, experimental setups with multiple cameras have to be developed. Development of such systems can largely benefit from a robust and easy to use camera calibration and 3D reconstruction toolbox. To map features tracked from multiple cameras using DLC to 3D world coordinates there exist OpenCV (@opencv) based implementations (@Nath2019, @Anipose). <br/>
+Markerless tracking is a crucial experimental requirement for behavioral studies conducted in many species in different environments. A recently developed toolbox called DeepLabCut (DLC) (@Mathis2018) leverages Artificial Neural Network (ANN) based computer vision to make precise markerless tracking possible for scientific experiments. DLC uses a deep convolutional neural network, ResNet (@He2016) pre-trained on ImageNet database (@Imagenet) and adapts it to make it applicable for behavioral tracking tasks. To track complex behaviors such as grasping with object interaction in 3D, experimental setups with multiple cameras have to be developed. Development of such systems can largely benefit from a robust and easy to use camera calibration and 3D reconstruction toolbox. To map features tracked from multiple cameras using DLC to 3D world coordinates there exist OpenCV (@opencv) based implementations (@Nath2019, @Anipose). <br/>
 
 In comparison to OpenCV, Matlab (The MathWorks Inc., Natick, Massachusetts) provides a feature-rich graphical user interface (GUI) for camera calibration which is an essential step for 3D reconstruction. The GUI provides visual feedback and helps quickly detect and reduce errors during the calibration process. However, existing functions in Matlab cannot be trivially used to reconstruct 3D behavior from 2D tracked data for more than two cameras. Our toolbox pose3d is implemented in Matlab for 3D reconstruction of features tracked in 2D using DLC or any other tracking software. It provides a semi-automated 3D reconstruction workflow that takes users through the entire process of camera calibration, undistortion, triangulation as well as post processing steps such as filtering to reduce outliers. pose3d also allows users to try different pre- and post-processing parameters that can be set by simply editing a configuration file and running the main function of the repository called main_pose3d. In addition, pose3d visualizes the results for every run of the main function and helps perform manual parameter tuning before saving final results. <br/>
 
@@ -41,5 +41,6 @@ Demo datasets provided in pose3d were acquired by moving a Rubik’s cube on a t
 
[…]
