### 2019-08-18 #1648 @cMadan
@SwathiSheshadri, it looks like the project is mainly a GUI utility that wraps around DeepLabCut and first-party Matlab functions. Do you think this is a fair assessment?
