### 2019-08-19 #1648 @SwathiSheshadri
@cMadan, thank you for taking time to edit our submission.
I agree with your assessment in that the focus of our submission with pose3d is to provide a user-friendly 3D reconstruction interface targeted mainly towards DeepLabCut (DLC) users. However, pose3d is not limited to DLC users and can be easily used with any other 2D tracking software as described in one of our demo examples.
I would also like to add that pose3d extends Matlab’s existing triangulation function designed for use with only 2 cameras to ‘n’ cameras and includes a description of the same in the Appendix.pdf document of our paper folder.
