### 2019-09-07 #1648 @SwathiSheshadri
@cMadan, I suggest @sreschechtko (Sasha Reschechtko) and @brittonsauerbrei (Britton Sauerbrei) as potential reviewers. These two researchers are actively interested in behavior tracking and would be well suited to review pose3d.
Additionally, I suggest the following reviewers from the JOSS list based on their programming language preference and area of expertise : @Kevin-Mattheus-Moerman, @fernandoandreotti
