### 2019-10-17 #1648 @Kevin-Mattheus-Moerman
@danasolav @cMadan will likely find a few more reviewers before the formal review starts. If you like you can start evaluating the software but you can also wait until the review issue is opened. There you'll have tickboxes to guide you through the process.
