### 2019-10-28 #1648 @SwathiSheshadri
Thank you all for your responses and for helping advance our submission towards review.
@cMadan, I make the following additional reviewer suggestions.
Additional JOSS reviewer suggestions:
@thomakra
@cpontonn
In addition, I request you to reconsider my earlier suggested reviewers @sreschechtko (Sasha Reschechtko) and @brittonsauerbrei (Britton Sauerbrei) as they would be best equipped to test our submission out. Please let me know if I have to contact them and request them to register as reviewers at JOSS.
