### 2019-10-30 49afa9e
Self defined color map code update
