### 2019-11-18 SwathiSheshadri/pose3d#3 @sreschechtko
[JOSS Review] Enhance plotting ability
The package would be more useful if it included ability to plot the 3D reconstruction in synchrony with at least one camera view. This functionality exists in the demo, but it is not implemented in the actual package, which is confusing. WIthout synchronized video it is difficult to verify the performance of the reconstruction.
