### 2019-11-18 SwathiSheshadri/pose3d#5 @sreschechtko
[JOSS Review] Enhance Troubleshooting section
Troubleshooting topics are vague and do not provide context for identifying what these errors look like at runtime. It would be much improved by showing demonstrations of the error output (either screenshots/gifs or explanations) of these failure modes so users can diagnose them.
