### 2019-11-18 #1849 @sreschechtko
I was able to install the package and run the demo as well as my own data. I have raised some issues (#[1-6](https://github.com/SwathiSheshadri/pose3d/issues)) which I ran into during this process and which I would like to see addressed before recommending the publication of this package in JOSS.
