### 2019-11-20 f1f8733
code changes to address issues raised by Sasha
