+++ Codes/testing_config_file.m
@@ -53,7 +53,7 @@ ncams = 5; %Edit this to match your current experiment (minimum allowed value is
 % There must be one path cell entry for primary camera DLC tracked data
 % copy and paste the full path to your csv files generated by DLC
 % this path can be anywhere that is readable by your user on the computer
-primary2D_datafullpath     = {'csvdatafullpath../Primary-Rubiks_1030000.csv'}; 
+primary2D_datafullpath     = {'/Users/ssheshadri/iMac_local/sandbox/DemoData/rubiks_primary-0000DeepCut_resnet50_RubiksCubeJul27shuffle1_600000.csv'}; 
 % If you are using other software use the csv files in Rubikscube_other2d folder as
 % reference to ensure the row and column entries are similar
 
@@ -67,17 +67,17 @@ primary2D_datafullpath     = {'csvdatafullpath../Primary-Rubiks_1030000.csv'};
 % in the variable calibvideos_secondary
 % For example : 2D tracked data from secondary1 should be from the same camera as the
 % secondary2 calib file
-secondary2D_datafullpath = [{'csvdatafullpath../Secondary1-Rubiks_1030000.csv'};...
-                            {'csvdatafullpath../Secondary2-Rubiks_1030000.csv'};...
-                            {'csvdatafullpath../Secondary3-Rubiks_1030000.csv'};...
-                            {'csvdatafullpath../Secondary4-Rubiks_1030000.csv'}];
+secondary2D_datafullpath = [{'/Users/ssheshadri/iMac_local/sandbox/DemoData/rubiks_secondary1-0000DeepCut_resnet50_RubiksCubeJul27shuffle1_600000.csv'};...
+                            {'/Users/ssheshadri/iMac_local/sandbox/DemoData/rubiks_secondary2-0000DeepCut_resnet50_RubiksCubeJul27shuffle1_600000.csv'};...
+                            {'/Users/ssheshadri/iMac_local/sandbox/DemoData/rubiks_secondary3-0000DeepCut_resnet50_RubiksCubeJul27shuffle1_600000.csv'};...
+                            {'/Users/ssheshadri/iMac_local/sandbox/DemoData/rubiks_secondary4-0000DeepCut_resnet50_RubiksCubeJul27shuffle1_600000.csv'}];
 
                         
 %% Calibration Specifics
 
 squareSize = 24; %%enter your checkboard square length in mm default is 24
 
-calib_videos = 1; %set this to zero if you have acquired calibration images and 1 if videos
+calib_videos = 0; %set this to zero if you have acquired calibration images and 1 if videos
 
 % Number of frames to extract from calibration videos to use for calibration
 % Matlab recommends 10 to 20. 
@@ -126,7 +126,7 @@ CalibrationImages_format = '.png';
 % calibration images from every camera pair are saved)
 % edit example path below to the path where you have saved your calibration
 % images
-folderwithpngs = '/Users/username/CalibFiles/';
+folderwithpngs = '/Users/ssheshadri/iMac_local/sandbox/TestExperiments1/Imagesforcalibration/';
 
 %% You are now done with fill up the main parameters!!! edit below for post-processing
 % (Default setting we use are below.)
@@ -139,10 +139,11 @@ run_undistort = 0; %1: run undistort, 0: do not undistort
 plotresults = 1; %plots resulting 3D coordinates at recorded fps
 
 
-path_to_2Dtracked_video = [exp_path '/' exp_name '/Videos/*.mp4']; %full path of the 2D tracked video you want to visualize alongside 3D tracked results
[…]
