### 2019-11-20 9c4158f
template configuration file updated
