+++ Codes/testing_config_file.m
@@ -1,232 +1,0 @@
-% Template config file for pose3d
-% Edit this file to enter your project specifics
-%
-% Copyright (c) 2019 Swathi Sheshadri
-% German Primate Center
-% swathishesh AT gmail DOT com
-%
-% If used in published work please see repository README.md for citation
-% and license information: https://github.com/SwathiSheshadri/pose3d
-
-%% Specify path and file name for pose3d to save your 3D reconstruction related files in
-% exp_path holds the name of the folder that gets created by pose3d for
-% your experiment. This folder will be created in your present working directory.
-% All experiments in this folder access the same calibration files
-% Therefore, everytime you move cameras or record from a different setup
-% change this name, otherwise you can keep this fixed
-
-exp_path = 'AllExperiments'; %./AllExperiments **To be different for every new camera configuration/experiment setup**
-
-% exp_name holds the name of the session of your experiment 
-% this is created as a folder withint exp_path directory by pose3d for your session specific data files
-
-exp_name = 'Experiment1'; % ./AllExperiments/Experiment1; **To be different for every new experiment**
-
-%Are you using DeepLabCut for 2D tracking?
-usingdlc = 1; %set to 0, if using other software
-
-%which mode of 3D reconstruction do you want to use
-% 1. 'all' (2D tracked data from all cameras is used for reconstruction
-% 2. 'bestpair' (2D tracked data from best camera pair is used for 3D
-% reconstruction for every time point and feature)
-% 3. 'avg' (3D reconstructed data averaged over pairs)
-% You can also select more than 1 mode and all specified moved will be saved in the specified ordered in Data3d.mat, for instance by setting modes_3drecon = {'all', 'bestpair'};
-modes_3drecon = {'all'}; %this is the recommended mode (other possible modes are {'bestpair'},{'avg'}};
-
-%% Enter your experiment specifics
-% Fill in the number of features you have tracked in 2D using DLC or any
-% other 2D tracking software (eg: for our demo nfeatures = 8;)
-% Change this to match your current experiment
-nfeatures = 8; %Edit this to match your current experiment
-
-% Fill in precise duration of recording in **seconds** (eg: for our demo rec_time = 10;)
-rec_time =  10; %Edit this to match your experiment recording time 
-
-% Fill in frames per second (eg: for our demo fps = 100;)
-fps =100; %Edit this to match your current experiment**
-
-% Fill in the overall number of cameras in your experiment (for our demo we had 5 cameras)
-% Change this to the number of cameras you have in your setup
-ncams = 5; %Edit this to match your current experiment (minimum allowed value is 2)
-
-%% Copy and paste your 2D tracked csv file paths with filenames
-% There must be one path cell entry for primary camera DLC tracked data
-% copy and paste the full path to your csv files generated by DLC
-% this path can be anywhere that is readable by your user on the computer
-primary2D_datafullpath     = {'/Users/ssheshadri/iMac_local/sandbox/DemoData/rubiks_primary-0000DeepCut_resnet50_RubiksCubeJul27shuffle1_600000.csv'}; 
-% If you are using other software use the csv files in Rubikscube_other2d folder as
-% reference to ensure the row and column entries are similar
-
-% There must be ncams-1 path cell entries for secondary cameras (we have 5 cameras of
-% which 1 is primary and remaining are secondary)
-% this path can be anywhere that is readable by your user on the computer
-%
-% **Important info** -- 
-%----------------------
[…]
