### 2019-11-24 446759e
updated document/codes as per Sasha's comments
