### 2019-11-25 d829da0
small doc updates
