+++ paper/paper.md
@@ -30,20 +30,18 @@ bibliography: paper.bib
 # Summary
 
 Markerless tracking is a crucial experimental requirement for behavioral studies conducted in many species in different environments. A recently developed toolbox called DeepLabCut (DLC) (@Mathis2018) leverages Artificial Neural Network (ANN) based computer vision to make precise markerless tracking possible for scientific experiments. DLC uses a deep convolutional neural network, ResNet (@He2016) pre-trained on ImageNet database (@Imagenet) and adapts it to make it applicable for behavioral tracking tasks. To track complex behaviors such as grasping with object interaction in 3D, experimental setups with multiple cameras have to be developed. Development of such systems can largely benefit from a robust and easy to use camera calibration and 3D reconstruction toolbox. The current version of DLC allows 3D reconstruction from pairs of cameras only (@Nath2019). <br/>
-We developed pose3d for 3D reconstruction of 2D features tracked from multiple cameras (n>=2). Our implementation of 3D reconstruction can be divided into two main steps namely stereo camera calibration and triangulation. Camera calibration refers to the estimation of intrinsic camera parameters namely focal length and optical center, extrinsic camera parameters namely the location of camera in world coordinates as well as the estimation of lens distortion coefficients. Additionally, in stereo camera calibration, one of the cameras is fixed as the primary and the position of the secondary camera with respect to the primary camera is estimated.  Typically, stereo camera calibration is performed by using checkerboard presented at varied angles to the pair of cameras being calibrated to establish the correspondence between 2D image coordinates and 3D world coordinates. Following this, camera parameters and distortion coefficients are estimated based on these correspondences. After stereo camera calibration is done for all secondary cameras with respect to a single selected primary camera, distortion in the measured 2D coordinates of the feature being tracked can be corrected and transformed into 3D world coordinates using triangulation. There exists other OpenCV (@opencv) based implementations in python particularly Anipose (@Anipose) which provide similar functionality as pose3d. However, in comparison to OpenCV based implementations, pose3d provides a user-friendly application by integrating with the feature-rich graphical user interface (GUI) for stereo camera calibration in Matlab (The MathWorks Inc., Natick, Massachusetts). A more detailed comparison between Anipose and pose3d is included in a subsequent section titled “Comparison of Anipose and pose3d”. <br/>
-To use pose3d for stereo camera calibration and 3D reconstruction the users need only to edit a configuration file to enter their experiment specific details. The configuration file is extensively commented to help users through the process of editing it. Running the main function of the repository ‘main_pose3d.m’ with user’s configuration file, creates the folder structure required for stereo camera calibration and 3D reconstruction automatically. Then, pose3d launches Matlab’s stereoCameraCalibrator GUI sequentially for every pair of primary and secondary cameras to estimate the parameters of the cameras in the user’s experimental setup. The GUI first parses through checkerboard frames recorded simultaneously from the two cameras being calibrated to select frames in which checkerboard can be detected in both cameras. At this point, the user can select the number of distortion co-efficients to be estimates. If the lens used in the experiment causes higher distortion of images, 3 distortion co-efficients can be used instead of the 2 (default) by simply selecting the radio-button on the GUI corresponding to it.  After this, clicking the calibrate button on the GUI, estimates camera parameters and relative positioning of secondary camera with respect to the primary. The GUI also displays reprojection errors across calibration frames allowing users to iteratively recalibrate after removing outliers (this can be done by right-clicking on the data browser of the GUI and selecting the option remove and recalibrate). After calibration is complete, session can be saved by clicking the save as button. This procedure is then repeated for all camera pairs to be calibrated. Pose3d prompts users throughout this procedure, providing messages on the next steps to be carried out. <br/>
[…]
