### 2019-11-25 50e6114
disp msg changed in extract frames fucntion
