+++ Codes/core_functions/extract_frames.m
@@ -44,7 +44,15 @@ try
         filename=strcat('frame',num2str(img),'.png');
         vidframe = read(vidfile, img);
         if ~mod(k,5)
-            disp(['Writing calibration image ' num2str(k) ' of ' num2str(frames_to_use) ' for ' cam_type ' of camera pair ' num2str(num_pair)])
+            if strcmp(cam_type,'Primary')
+                if contains(dest_path,'PrimarySecondary')
+                    disp(['Writing calibration image ' num2str(k) ' of ' num2str(frames_to_use) ' for ' cam_type ' of camera pair ' num2str(num_pair)])
+                else
+                    disp(['Writing calibration image ' num2str(k) ' of ' num2str(frames_to_use) ' for ' cam_type])
+                end
+            else
+                disp(['Writing calibration image ' num2str(k) ' of ' num2str(frames_to_use) ' for ' cam_type ' of camera pair ' num2str(num_pair)])
+            end
         end
         imwrite(vidframe,[dest_path filename]);
         k=k+1;
