+++ paper/paper.md
@@ -43,7 +43,7 @@ pose3d is a 3D reconstruction workflow integrated closely with functions from Ma
 # Comparison of pose3d and anipose
 Functionally, pose3d and anipose provide users with pre- and post-processing tools for 3D reconstruction including the process of camera calibration. The main difference is that pose3d provides a user-friendly GUI utility which is missing in Anipose. With pose3d by entering experiment details in configuration file and running a single function in Matlab namely ‘main_pose3d.m’ the user can perform the entire procedure of 3D reconstruction. Furthermore, integration of pose3d with stereoCameraCalibrator GUI in Matlab makes the task of camera calibration very intuitive. This allows users to estimate camera and lens distortion parameters, select outlier frames from calibration with the click of a button. 
 On the other hand, anipose uses OpenCV functions which can be quite difficult for non-experts to get into. For instance, with pose3d, due to its integration with Matlab’s stereoCameraCalibrator GUI, all pairs of images used for camera calibration are visualized along with their reprojection errors allowing easy identification and removal of images with larger error values. With OpenCV functions, users get error messages that are often quite technical in nature and can take longer for non-experts to perform root-cause analysis. Lastly, with pose3d demo datasets including example calibration images are included along with codes to run 3D reconstruction on demo datasets. Furthermore, tutorial videos illustrating the usage of pose3d are included. <br/>
-Some other minor similarities and differences between anipose and pose3d are listed in the following section. First, both anipose and pose3d allow users to have the same set of calibration files across sessions or have a separate set per session. This can be decided by the user depending on if the user wants to have the same calibration files across sessions or not. This is possible in pose3d by having the same experiment path (assigned to variable exp_path) and changing only the experiment name (assigned to variable exp_name) in the configuration file when the same calibration files are to be used across sessions. With anipose this is implemented by having a hierarchy for the selection of calibration videos placed in different folders. Second, the triangulation approach used is the same in the two packages. However, with pose3d for comparison purposes we have included 3 modes of 3D reconstruction namely 'all', 'avg' and 'best-pair'. Of these three modes 'all' corresponds to the one implemented in anipose. On our demo dataset, we have shown 'all' mode to give the lowest on average error and is the recommended mode. Third, anipose allows for camera calibration using charuco boards, checker boards or aruco boards of which it recommends the usage of charuco and checker boards. Since pose3d integrates with the camera calibration GUI from MATLAB, only checker board is used for calibration. <br/>
+Some other minor similarities and differences between anipose and pose3d include the following. First, both anipose and pose3d allow users to have the same set of calibration files across sessions or have a separate set per session. This can be decided by the user depending on if the user wants to have the same calibration files across sessions or not. This is possible in pose3d by having the same experiment path (assigned to variable exp_path) and changing only the experiment name (assigned to variable exp_name) in the configuration file when the same calibration files are to be used across sessions. With anipose this is implemented by having a hierarchy for the selection of calibration videos placed in different folders. Second, the triangulation approach used is the same in the two packages. However, with pose3d for comparison purposes we have included 3 modes of 3D reconstruction namely 'all', 'avg' and 'best-pair'. Of these three modes 'all' corresponds to the one implemented in anipose. On our demo dataset, we have shown 'all' mode to give the lowest on average error and is the recommended mode. Third, anipose allows for camera calibration using charuco boards, checker boards or aruco boards of which it recommends the usage of charuco and checker boards. Since pose3d integrates with the camera calibration GUI from MATLAB, only checker board is used for calibration. <br/>
 For further technical reading on details of triangulation for 3D reconstruction please refer to our [supporting document](Appendix.pdf) <br/>
 # References
 
