### 2019-12-02 f8f7430
doc edits from Hans comments
