+++ paper/paper.md
@@ -29,21 +29,20 @@ bibliography: paper.bib
 
 # Summary
 
-Markerless tracking is a crucial experimental requirement for behavioral studies conducted in many species in different environments. A recently developed toolbox called DeepLabCut (DLC) (@Mathis2018) leverages Artificial Neural Network (ANN) based computer vision to make precise markerless tracking possible for scientific experiments. DLC uses a deep convolutional neural network, ResNet (@He2016) pre-trained on ImageNet database (@Imagenet) and adapts it to make it applicable for behavioral tracking tasks. To track complex behaviors such as grasping with object interaction in 3D, experimental setups with multiple cameras have to be developed. Development of such systems can largely benefit from a robust and easy to use camera calibration and 3D reconstruction toolbox. The current version of DLC allows 3D reconstruction of features tracked in 2D from pairs of cameras only (@Nath2019). This solution is not sufficient when behavior is tracked with more than 2 cameras. Furthermore, for recording conditions that might result in noisy 2D tracking such as low light and low resolution, the accuracy in the tracked 3D coordinate can be improved by using data from more than two cameras. <br/>
+Markerless tracking is a crucial experimental requirement for behavioral studies conducted in many species in different environments. A recently developed toolbox called DeepLabCut (DLC) (@Mathis2018) leverages Artificial Neural Network (ANN) based computer vision to make precise markerless tracking possible for scientific experiments. DLC uses a deep convolutional neural network, ResNet (@He2016) pre-trained on ImageNet database (@Imagenet) and adapts it to make it applicable for behavioral tracking tasks. To track complex behaviors such as grasping with object interaction in 3D, experimental setups with multiple cameras are required. Such systems can largely benefit from a robust and easy to use camera calibration and 3D reconstruction toolbox. The current version of DLC allows 3D reconstruction of features tracked in 2D from pairs of cameras only (@Nath2019) and is not sufficient when behavior is tracked with more than 2 cameras. Furthermore, for noisy 2D tracking conditions, such as low light or low resolution, the accuracy of tracked 3D coordinates can be improved by evaluating data from more than two cameras.  <br/>
 To facilitate 3D tracking of complex behaviors requiring multiple cameras (n>=2), we developed pose3d: a Matlab (The MathWorks Inc., Natick, Massachusetts) implementation for 3D reconstruction. pose3d can be divided into two main parts: camera calibration with pairwise estimation of relative camera positions to each other and 3D reconstruction with triangulation performed over 2D feature coordinates tracked from all cameras. Camera calibration refers to the estimation of intrinsic camera parameters including the level of magnification captured by the focal length and the position in space given by the optical center of the lens, extrinsic camera parameters including the location of camera in 3D coordinates as well as the distortion in the camera image caused by the lens. In stereo camera calibration, one of the cameras is fixed as the primary camera and the position of the secondary camera with respect to the primary camera is estimated and given as an additional parameter. Stereo camera calibration in pose3d is performed by using checkerboard presented at varied angles to the pair of cameras being calibrated to establish the correspondence between 2D image coordinates and 3D coordinates. Following calibration, the 2D feature tracked across cameras is transformed into 3D coordinates using triangulation.  pose3d implements triangulation by minimizing the distance of the estimated 3D feature coordinate from the lines passing through the focal center and the image plane of the cameras simultaneously. <br/>
 # pose3d usage and features
[…]
