### 2019-12-02 0894d08
minor doc edits
