### 2019-12-02 SwathiSheshadri/pose3d#3 @SwathiSheshadri
A very good point, thank you! We have updated the simple_3Ddata_plotter.m file to now show 3D reconstructed data alongside user selected 2D tracked video. The path to the video the user wishes to visualize alongside the 3D reconstructed results has to be entered in the config file. Visualization of more than one 2D tracked videos alongside 3D reconstructed data is also possible using pose3d. This feature is included as a helper function named make_illustrative_movie.m and can be used optionally.
