### 2019-12-02 SwathiSheshadri/pose3d#5 @SwathiSheshadri
To address this issue, we have included in the section renamed ‘Video tutorial and troubleshooting’ video tutorials on using pose3d and added elaborated explanations to the trouble shooting section.
