### 2019-12-02 SwathiSheshadri/pose3d#7 @danasolav
[JOSS Review] Enhance csv file instructions
The instructions should clearly explain how to create the csv file in case the 2D points are not generated using DLC. In my opinion, asking the user to use a demo file as a reference and learn from it is insufficient.
