### 2019-12-02 #1849 @danasolav
I have installed and ran the demo successfully. I would like to run my own data which was not acquired using DLC. Therefore I have raised an issue ([#7](https://github.com/SwathiSheshadri/pose3d/issues/7)) regarding the instructions for creating the csv file which is used as input.
