### 2019-12-03 35a3b9e
update to addressing Dana's issue # 7
