+++ Codes/template_config_file.m
@@ -22,7 +22,9 @@ exp_path = 'AllExperiments'; %./AllExperiments **To be different for every new c
 
 exp_name = 'Experiment1'; % ./AllExperiments/Experiment1; **To be different for every new experiment**
 
-%Are you using DeepLabCut for 2D tracking?
+%If you are using DeepLabCut for 2D tracking or if your 2D tracking software also
+%provides likelihood values ranging between 0 and 1 similar to DLC set this
+%variable to 1; otherwise set this to 0
 usingdlc = 1; %set to 0, if using other software
 
 %which mode of 3D reconstruction do you want to use
