### 2019-12-03 7df5307
minor edits addressing Dana's issue # 7
