### 2019-12-03 #1849 @SwathiSheshadri
@sreschechtko, please see our responses on the issues you raised at pose3d (#[1-6](https://github.com/SwathiSheshadri/pose3d/issues)). We have addressed the issues and have updated the repository to reflect the changes. Please use the latest release of pose3d v2.0.0 to see the changes done to pose3d as per your comments. We found your feedback and detailed comments very helpful for us in refining our repository for public use and we thank you very much again for your time and effort and look forward to hearing from you!
