### 2019-12-03 SwathiSheshadri/pose3d#7 @SwathiSheshadri
Thank you for pointing this out! The readme file is now up-to-date with details on how to prepare 2D tracked data for pose3d when using tracking software other than DLC. Specifically, we have updated the section 'Data requirements for pose3d' on the readme file to provide clear guidelines to users.
