### 2019-12-03 #1849 @SwathiSheshadri
@danasolav, we developed pose3d using our 2D tracked data from DLC and we are very happy to hear that you have 2D tracked data outside of DLC framework as this will help us get your  perspective on our code for this particular use-case. In our latest release of pose3d (v2.0.0), we have updated the readme file (addressing issue [#7](https://github.com/SwathiSheshadri/pose3d/issues/7) to include details on preparing the .csv files required for pose3d when using a different software for 2D tracking. Looking forward to hearing more about your experience with pose3d!
