### 2019-12-03 SwathiSheshadri/pose3d#7 @danasolav
Thank you for the prompt response. I was able to prepare the csv and run it, but encountered the following error:
Struct contents reference from a non-struct array object.
```
Error in load_preprocess_2d_data>load_otherdata (line 207)
    data2d = data2d.data;
```
The reason is that when my csv file is imported using importdata, it does not create a structure with a data field, it creates a matrix (possibly because my csv file does not contain any text or headers).

In addition, I have a minor comment about the way these guidelines appear in the readme file- it seems as if it is a part of point number 3, which is about calibration. I think the instructions for creating the csv file should be given in a new bullet point.
