### 2019-12-03 SwathiSheshadri/pose3d#7 @SwathiSheshadri
Thank you for your quick response and for giving me the details of the error. The latest release of the repository v2.0.0 includes a check to see if the variable is a structure and references the data field afterwards. I included this check in the code, since I anticipated that some users might have text fields in their .csv files and some others not. Could you please let me know if this error still exists when running the latest code available at the repository (released as version v2.0.0)?
