### 2019-12-03 1044c0c
Adding feature to vizualize 3D data with 2D tracked images(Dana Issue #8)
