### 2019-12-03 84c4285
Minor code edit(Dana Issue #8)
