### 2019-12-04 78b9d61
documentation edits addressing Dana's issue # 8
