### 2019-12-04 d2691d5
documentation edits addressing Dana's issue #10
