### 2019-12-05 753932c
minor code improvements after testing
