### 2019-12-09 SwathiSheshadri/pose3d#5 @sreschechtko
Thanks, I think the video tutorials are helpful. On a high resolution display, I had a hard time following the tutorial because the text was very small.  I  will close this issue but would recommend either adding more explanatory subtitles, using magnification, or including a separate text file with the tasks you carry out and timestamps when you are doing them.
