### 2019-12-09 SwathiSheshadri/pose3d#3 @sreschechtko
Thank you for adding this functionality. I think it would be better implemented by having the option to input a path to an untracked video and then to plot the 2d points on this video since the program already has access to the tracked points in 2d. At present, if someone does not have 2d tracked videos saved, they would have to go back into another application to produce these videos.

Additionally, I think the error thrown by the plotter when no video (or images) is present but plotting tracked video option is on is difficult to interpret because it just says the file doesn't exist at this path and requires multiple layers of trackback. Please raise a more specific error.
