### 2019-12-09 #1849 @sreschechtko
Thank you for responding to my review issues. I have closed the majority of issues because I think they are satisfactorily addressed, although I had a few minor comments on [5](https://github.com/SwathiSheshadri/pose3d/issues/5) and [7](https://github.com/SwathiSheshadri/pose3d/issues/7).

The only issue I raised which I am leaving open is [3](https://github.com/SwathiSheshadri/pose3d/issues/3) for plotting 2D data with the 3D reconstruction. I think pose3d should have the option to plot the 2d points (which it already has access to) on an un-annotated video since annotated video is not necessarily commonly saved output. I would also like to a see more descriptive error message in the plotter if the video is not available but the option to plot it has been selected.
