### 2019-12-10 abc5e95
Updates to improve plotting functions issue #3,number of user inputs needed in config file reduced for further ease of use
