### 2019-12-10 8a8a593
code edits to improve plotting functions #3 and reduce amount of user input needed in config file
