+++ Codes/helper_functions/make_illustrative_movie.m
@@ -4,9 +4,6 @@
 % Compute 3D data first before running this code for your experiment
 % This function plots 3D reconstruction data alongside 2D tracking performed from all cameras
 % In addition this function makes a movie of the same
-% 
-% Before running this function
-% Video files must be present at [exp_path '/' exp_name '/Videos/']
 %
 % Copyright (c) 2019 Swathi Sheshadri
 % German Primate Center
@@ -20,36 +17,53 @@ clear
 close all
 clc
 
-%Take initialized values from config file
-template_config_file %call your config file here
+%%Take initialized values from config file
+testing_config_file %call your config file here
 
-load([exp_path '/' exp_name '/Data3d/Data3d.mat'],'coords3dall')
+load([exp_path '/' exp_name '/Data3d/Data3d.mat'],'coords3d')
+coords3dall = coords3d;
 
 %Code to visualize 2d videos with 3d reconstruction results
-figure('units','normalized','outerposition',[0 0 1 1])
+figure('units','normalized','outerposition',[0 0 1 0.5])
 colorclass = colormap(jet); %jet is default in DLC to-date
-color_map_self=colorclass(1:nfeatures:64,:);
+color_map_self=colorclass(ceil(linspace(1,64,nfeatures)),:);
 
-%Place your DLC labelled video files in the project under Videos folder
 
-v = dir([exp_path '/' exp_name '/Videos/*.mp4']); %Place your DLC labelled video files at this location
-cam = cell(ncams,1);
-for icams = 1:ncams
-    cam{icams,1} = VideoReader([v(icams,1).folder '/' v(icams,1).name]);
+
+%path to the videos you want to visualize next to 3D tracked data 
+tracked_videos = [{'Fullpath/Video1.mp4';};...    
+                  {'Fullpath/Video2.mp4';};...   
+                  {'Fullpath/Video3.mp4';}];
+
+%path 2D tracked csv files of the videos provided above
+path_to_csv = [{'Fullpath/csvfileofVideo1.csv'};... %provide csv file order should
+               {'Fullpath/csvfileofVideo2.csv'};... %correspond to the movie file order 
+               {'Fullpath/csvfileofVideo3.csv'};];
+                          
+ncams_to_display = size(tracked_videos,1);              
+cam = cell(ncams_to_display ,1);
+DataAll = nan(nframes,nfeatures*2,ncams_to_display ) ;
+
+for icams = 1:ncams_to_display
+    
+if usingdlc
+    [data2d,data2dllh,flag_mis] = load_dlcdata(path_to_csv{icams},nframes,nfeatures,flag_mis);
+else
+    [data2d,flag_mis] = load_otherdata(path_to_csv{icams},nframes,nfeatures,flag_mis);
+end
+    DataAll(:,:,icams) = data2d;
+    cam{icams,1} = VideoReader([tracked_videos{icams,1}]);
 end
 
 %to make video
-vidfile = VideoWriter([exp_path '/' exp_name '/Videos/' exp_name '2D_3Dmovie.mp4'],'MPEG-4');
+if ~exist([exp_path '/' exp_name '/Videos/'],'dir')
+    mkdir([exp_path '/' exp_name '/Videos/'])
+end
+vidfile = VideoWriter([exp_path '/' exp_name '/Videos/' exp_name '2D_3Dmovie.avi']);
 vidfile.FrameRate = fps;
 open(vidfile)
-
-if ~mod(size(cam,1),2)
-    subplot_size = size(cam,1);
-else
-    subplot_size = size(cam,1)+1;
-end
[…]
