+++ Codes/template_config_file.m
@@ -41,14 +41,14 @@ modes_3drecon = {'all'}; %this is the recommended mode (other possible modes are
 % Change this to match your current experiment
 nfeatures = 8; %Edit this to match your current experiment
 
-% Fill in precise duration of recording in **seconds** (eg: for our demo rec_time = 10;)
-%If you have acquired images and no videos during your experiment rec_time and fps values 
-%are not important (you can fill any number into them or leave blank). Make sure to enter the right value to variable
-%nframes in the section of this config file called "Edit below for
-%post-processing and data visualization"
-rec_time =  10; %Edit this to match your experiment recording time 
-% Fill in frames per second (eg: for our demo fps = 100;)
-fps =100; %Edit this to match your current experiment**
+% % Fill in precise duration of recording in **seconds** (eg: for our demo rec_time = 10;)
+% %If you have acquired images and no videos during your experiment rec_time and fps values 
+% %are not important (you can fill any number into them or leave blank). Make sure to enter the right value to variable
+% %nframes in the section of this config file called "Edit below for
+% %post-processing and data visualization"
+% rec_time =  10; %Edit this to match your experiment recording time 
+% % Fill in frames per second (eg: for our demo fps = 100;)
+% fps =100; %Edit this to match your current experiment**
 
 % Fill in the overall number of cameras in your experiment (for our demo we had 5 cameras)
 % Change this to the number of cameras you have in your setup
@@ -142,15 +142,23 @@ run_undistort = 0; %1: run undistort, 0: do not undistort
 % values so only a few example frames are vizualized while saving everything)
 plotresults = 1; %plots resulting 3D coordinates at recorded fps
 
-have2Dtrackedvideos = 1; %set this to 0 if you have 2D tracked images instead of videos
+%setting both have2Dtrackedvideos and have2Dtrackedimages to 0, lets you
+%visualize only the 3D reconstructed data. 
+%It is recommended to set have2Dtrackedvideos and have2Dtrackedimages to 0, if your video/images size is larger than your working memory use
+%make_illustrative_movie.m function in the repository for 3D reconstruction
+%visualization alongside 2D video/images
+have2Dtrackedvideos = 1; %set this to 1 if you have 2D tracked videos
+have2Dtrackedimages = 0; %set this to 1 if you have 2D tracked images instead of videos
 
-%this variable is used when have2Dtrackedvideos is 1
-path_to_2Dtracked_video = '.../*.mp4'; %full path of the 2D tracked video you want to visualize alongside 3D tracked results
+%this variable is used when have2Dtrackedvideos is 1 (provide path to
+%primary camera video here)
+path_to_2Dtracked_video = '.../*.mp4'; %full path of the 2D video you have tracked from **primary camera** to visualize alongside 3D tracked results
 
 %below 2 variables are only needed when you have 2D tracked images instead
-%of videos
-path2Dtrackedimages_folder = '/Users/username/2DTrackedImageFiles/'; %full path to the folder holding 2D tracked images
+%of videos (provide path to images tracked by primary camera
+path2Dtrackedimages_folder = '/Users/username/2DTrackedImageFiles/'; %full path to the folder holding 2D tracked images from **primary camera**
 format_of_images = '.png'; %can also be '.tif','.jpeg' or any other format supported by imread function in matlab
+
 
 color_bw = 1; % 1 if video/images of 2D tracked data is in color and 0 for black and white
 
@@ -168,14 +176,25 @@ drawline = [ 1 2; 2 3; 3 4; 4 1; 5 6;6 7;...
[…]
