### 2019-12-10 363f7a1
minor code edits
