### 2019-12-11 5c6c742
updated plotting fxn bug fix
