### 2019-12-11 4085f70
updated helper fxn make illustrative movie to allow users to plot 2D coordinates onto unlabelled movies/images
