+++ paper/paper.md
@@ -41,7 +41,7 @@ The key difference between the two examples is as follows. DLC, in addition to 2
 # Comparison of pose3d to existing Matlab functions
 pose3d is a 3D reconstruction workflow integrated closely with functions from Matlab’s computer vision toolbox. This workflow fixes one of the cameras in the experimental setup as primary which is crucial to ensure that the data points are within the same coordinate system across camera pairs. The stereo camera calibration part of our implementation can be considered a wrapper around functions from Matlab’s computer vision toolbox, however, the triangulation function included with pose3d extends the existing triangulation function in Matlab that operates only for 2 cameras. <br/>
 # Comparison of pose3d and anipose
-Functionally, pose3d and anipose provide users with pre- and post-processing tools for 3D reconstruction including the process of camera calibration. The main difference is that pose3d provides a user-friendly GUI utility which is missing in Anipose. With pose3d, by entering the experimental details in a configuration file and running a single Matlab function (‘main_pose3d.m’), the user can perform the entire procedure of 3D reconstruction. Furthermore, the integration of pose3d with the stereoCameraCalibrator GUI in Matlab makes the task of camera calibration very intuitive. This allows users to estimate camera and lens distortion parameters, select outlier frames from calibration with the click of a button. On the other hand, anipose uses OpenCV functions that can be quite difficult to use for non-experts. For instance, with pose3d, due to its integration with Matlab’s stereoCameraCalibrator GUI, all pairs of images used for camera calibration are visualized along with their reprojection errors, which allows easy identification and removal of images with larger error values. With OpenCV functions, users get warning/error messages that are often quite technical in nature and can take longer for non-experts to perform root-cause analysis. Lastly, pose3d provides demo datasets including example calibration images along with code to run 3D reconstructions on demo datasets as well as video tutorials to illustrate its operation.  <br/>
+Functionally, pose3d and anipose provide users with pre- and post-processing tools for 3D reconstruction including the process of camera calibration. The main difference is that pose3d provides a user-friendly GUI utility which is missing in Anipose. With pose3d, by entering the experimental details in a configuration file and running a single Matlab function (‘main_pose3d.m’), the user can perform the entire procedure of 3D reconstruction. Furthermore, the integration of pose3d with the stereoCameraCalibrator GUI in Matlab makes the task of camera calibration very intuitive. This allows users to estimate camera and lens distortion parameters, select outlier frames from calibration with the click of a button. On the other hand, anipose uses OpenCV functions that can be quite difficult to use for non-experts. For instance, with pose3d, due to its integration with Matlab’s stereoCameraCalibrator GUI, all pairs of images used for camera calibration are visualized along with their reprojection errors, which allows easy identification and removal of images with larger error values. With OpenCV functions, users get warning/error messages that are often quite technical in nature and can take longer for non-experts to perform root-cause analysis. In addition, pose3d provides demo datasets including example calibration images along with code to run 3D reconstructions on demo datasets as well as video tutorials to illustrate its operation. Lastly, error calculation routine is included in pose3d to quantify deviation of the reconstructed data from ground truth which is not provided in the current version of Anipose. For example, in the demo dataset using Rubik’s cube the 12 edges of the cubes were physically measured and compared against the reconstructed edge lengths obtained after running pose3d. Similarly, fixed lengths in the user’s experimental setups that are tracked can be compared against the measured values using the error calculation routine of pose3d.<br/>
[…]
