### 2019-12-11 #1849 @SwathiSheshadri
Thank you for closing the issues @sreschechtko
I have addressed the open issue #3 as well as the minor comments in the latest release of pose3d (v2.1.0).
The plotting function has been updated to allow feature annotation along with plotting video frames/images used for 2D tracking recorded from primary camera. Alternatively, make_illustrative_movie.m helper function is now updated to allow users to flexibly select from any number of perspectives they want to visualize alongside 3D reconstructed results. Furthermore, error handling in the config file has been upgraded to deal with missing video/image path entries in the config file. I have also removed the requirement of user input for frame rate and duration of recording parameters in the config file to make editing the config file easier, thereby, less error prone.
I again thank you both (@danasolav and @sreschechtko) for your time and very helpful feedback and would be very happy to hear back from you.
