### 2019-12-11 SwathiSheshadri/pose3d#3 @SwathiSheshadri
Thank you for this feature suggestion. The plotting function has been updated to allow feature annotation along with plotting video frames/images used for 2D tracking recorded from primary camera. Since, primary camera is typically positioned to have the best view, pose3d plots movies or images recorded from primary camera next to 3D reconstructed results. I took this approach to reduce the number of user inputs required when filling the config file. Alternatively, make_illustrative_movie.m helper function is now updated to allow users to flexibly select any number of perspectives they want to visualize alongside 3D reconstructed results. Furthermore, error handling in the config file has been upgraded to deal with missing video/image path entries in the config file.
