### 2019-12-16 SwathiSheshadri/pose3d#3 @sreschechtko
Thanks for the updates; could you explain in more detail how make_illustrative_movie is supposed to be used? Is it intended to be a standalone function or is it called by another helper?
