### 2019-12-16 SwathiSheshadri/pose3d#3 @SwathiSheshadri
make_illustrative_movie is a standalone function which can be run to visualize 3D reconstructed results alongside any number of 2D videos recorded during the experiment. This can be used by editing the make_illustrative_movie.m function to include the config file that was used to do 3D reconstruction and then providing paths to the videos and .csv files holding 2D tracked videos for the respective videos. Successful execution of make_illustrative_movie.m function, plots all frames from the selected videos, alongside 3D reconstructed results and saves the results in .avi format in a subfolder called Videos in your experiment folder.
