### 2019-12-18 aaf5b08
Hardcoded axis limits in make_illustrative_movie.m function updated to automatic limit selection
