### 2019-12-18 SwathiSheshadri/pose3d#3 @SwathiSheshadri
Thanks for following up on this issue Sasha. I was looking into make_illustrative_movie.m function and found that I had hardcoded the axis limits! Since it worked for our data, it passed all the tests I did! This function is now updated to automatically set axis limits based on the maximum and minimum values in the dataset it is used on. Please download the latest version of make_illustrative_movie.m function from the repository and let me know if you have any other question/suggestion for this function.
