### 2019-12-21 SwathiSheshadri/pose3d#3 @sreschechtko
Thanks for the update. I got it to work! I feel like it would make more sense to just automatically generate the video using this script instead of the simple 3D data plotter, but maybe you're of a different opinion 😃
