### 2019-12-21 #1849 @sreschechtko
Hi All, I have closed all of the issues I opened on the main repository and am happy to recommend the package/article for publication in JOSS.
