### 2019-12-21 SwathiSheshadri/pose3d#3 @SwathiSheshadri
Thanks for verifying that the update works! Super happy to hear back from you :)). I programmed simple 3D data plotter for quick visualization purposes,  since the movie generation with make_illustrative_movie.m can take a long while especially if users have high data volumes coming from a large number of cameras, high resolution images, higher frame rate etc,. Also, from our recording setup we tried out the visualization from 5 cameras, and it was too overwhelming to view all 5 2D movies alongside 3D reconstruction results at once.
I will add a pointer in the readme file to direct users to try out the make_illustrative_movie.m file after running the main function and verifying the results. Thanks again so much for your very helpful feedback and suggestions :)).
