### 2019-12-21 2fcf223
comments/readme update
