### 2019-12-21 #1849 @SwathiSheshadri
Thank you for closing all the issues @danasolav and @sreschechtko. Your feedback was immensely valuable towards making pose3d better for public use and we are very grateful to both of you for your time and effort! @cMadan, we are very thankful to you and the JOSS team for taking time to edit our submission. We request you to use the latest release of pose3d - version v2.2.0, for further processing at JOSS and look forward to hearing from you about the next steps to take for our submission!
