### 2019-12-25 cbb876c
Formatting edits for paper doc
