### 2019-12-25 eb13e8e
paper edits
