### 2019-12-25 #1849 @SwathiSheshadri
@cMadan, thank you for your quick response! I have updated the archive name in Zenodo to match the name of our JOSS paper. Also, I have proofread the PDF file and have updated the paper.md file to have improved formatting for PDF. During the process, I noticed that the link in the paper PDF file to the archive is broken, please help with fixing the same.
