### 2019-12-25 #1849 @cMadan
@SwathiSheshadri, I'm not sure why that isn't working, but will look into it.

@openjournals/joss-eics, it looks like the archive link in the compiled PDF isn't working. Any ideas why (or how we might otherwise resolve this)?
