### 2019-12-26 #1849 @Kevin-Mattheus-Moerman
@cMadan
> @openjournals/joss-eics, it looks like the archive link in the compiled PDF isn't working. Any ideas why (or how we might otherwise resolve this)?

Those links should be fixed in the final accepted version. Some automatically generated links are not active in these proofs (all author generated links should work though).
