### 2019-12-29 #1849 @arfon
> @openjournals/joss-eics, it looks like the archive link in the compiled PDF isn't working. Any ideas why (or how we might otherwise resolve this)?

Yeah, sorry about this. That's undocumented behaviour and very confusing...
