### 2019-12-29 #1849 @arfon
@cMadan - as an editor you are able to ask Whedon to make final proofs of the PDF with `@whedon accept` and these PDFs should have working links.
