### 2020-01-06 #1849 @danielskatz
@SwathiSheshadri  - please fix the cases (upper vs lower) in the bib entries - use {}s to protect cases where needed.
