### 2020-01-06 #1849 @danielskatz
@cMadan - Other than the case issue in the bib, is this otherwise ready to accept?
