### 2020-01-06 6969a92
case fixes in bibliography
