+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @article {Nath2019,
-  title = {Using DeepLabCut for 3D markerless pose estimation across species and behaviors},
+  title = {Using {DeepLabCut} for 3D markerless pose estimation across species and behaviors},
   journal = {Nature Protocols},
   year = {2019},
   abstract = {Noninvasive behavioral tracking of animals during experiments is critical to many scientific pursuits. Extracting the poses of animals without using markers is often essential to measuring behavioral effects in biomechanics, genetics, ethology, and neuroscience. However, extracting detailed poses without markers in dynamically changing backgrounds has been challenging. We recently introduced an open-source toolbox called DeepLabCut that builds on a state-of-the-art human pose-estimation algorithm to allow a user to train a deep neural network with limited training data to precisely track user-defined features that match human labeling accuracy. Here, we provide an updated toolbox, developed as a Python package, that includes new features such as graphical user interfaces (GUIs), performance improvements, and active-learning-based network refinement. We provide a step-by-step procedure for using DeepLabCut that guides the user in creating a tailored, reusable analysis pipeline with a graphical processing unit (GPU) in 1{\textendash}12 h (depending on frame size). Additionally, we provide Docker environments and Jupyter Notebooks that can be run on cloud resources such as Google Colaboratory.},
@@ -23,7 +23,7 @@
 @inproceedings{He2016,
   title={Deep residual learning for image recognition},
   author={He, Kaiming and Zhang, Xiangyu and Ren, Shaoqing and Sun, Jian},
-  booktitle={Proceedings of the IEEE conference on computer vision and pattern recognition},
+  booktitle={Proceedings of the {IEEE} conference on computer vision and pattern recognition},
   pages={770--778},
   year={2016},
   doi={10.1109/CVPR.2016.90}
@@ -31,7 +31,7 @@
 
 @inproceedings{Imagenet,
   author = {Deng, J. and Dong, W. and Socher, R. and Li, L.-J. and Li, K. and Fei-Fei, L.},
-  title = {ImageNet: A Large-Scale Hierarchical Image Database},
+  title = {ImageNet: A large-scale hierarchical image database},
   booktitle = {CVPR09},
   year = {2009},
   doi = {10.1109/CVPR.2009.5206848}
@@ -49,7 +49,7 @@
 }
 
 @misc{opencv,
-  title={Open Source Computer Vision Library},
+  title={{Open} {Source} {Computer} {Vision} {Library}},
   year={2015},
   howpublished = {\url{https://opencv.org}}
 }
