### 2020-01-06 #1849 @SwathiSheshadri
@danielskatz, thank you for your response and input on the case issue in the .bib file. It is now updated.
